requirements.txt
google-api-python-client
python-dotenv
google-generativeai
gradio
transformers
moviepy
soundfile
datasets
chardet